Sync OpenBSD patchset 1114:
Instead of passing stdin/stdout/stderr file descriptors over imsg and
handling them in the server, handle them in the client and pass buffers
over imsg. This is much tidier for some upcoming changes and the
performance hit isn't critical.

The tty fd is still passed to the server as before.

This bumps the tmux protocol version so new clients and old servers are
incompatible.


git-svn-id: https://tmux.svn.sourceforge.net/svnroot/tmux/trunk@2800 303bd1e2-03bb-47f1-b221-2a0928954661